refactor: HttpVersion 패키지 조정 및 불필요 메서드 제거
